fix duplicate bundle size with monaco vim
