+++ src/components/MonacoEditor/monaco-editor-types.ts
@@ -113,4 +113,6 @@ export interface EditorProps {
   onMount?: OnMount;
 
   onBeforeDispose?: Function;
+
+  vim?: boolean;
 }
